Split diagnostics gate tests and update Learn Zen
