+++ tests/integration/cli_build/frontend_json/diagnostics_gate_golden.rs
@@ -1,6 +1,13 @@
 use std::path::Path;
 use std::process::Command;
 
+#[path = "diagnostics_gate_golden/actor_and_imports.rs"]
+mod actor_and_imports;
+#[path = "diagnostics_gate_golden/effects_text_control.rs"]
+mod effects_text_control;
+#[path = "diagnostics_gate_golden/intrinsics.rs"]
+mod intrinsics;
+
 fn fixture(path: &str) -> std::path::PathBuf {
     Path::new(env!("CARGO_MANIFEST_DIR")).join(path)
 }
@@ -51,323 +58,4 @@ fn assert_gate_diagnostics_golden(
         .unwrap_or_else(|err| panic!("read {}: {err}", expected_path.display()));
 
     assert_eq!(normalized.trim(), expected.trim());
-}
-
-#[test]
-fn emit_json_diagnostics_typed_allocator_effect_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "typed_allocator_effect_gate.zen",
-        r#"
-main = (allocator: Allocator<i32, Sync>) void { }
-"#,
-        "typed allocator effect gate",
-        "typed allocator/effect gate should emit one feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_typed_allocator_effect_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_sync_effect_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "sync_effect_gate.zen",
-        r#"
-main = (mode: Sync) void { }
-"#,
-        "sync effect gate",
-        "sync effect gate should emit one Sync/Async feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_sync_effect_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_async_effect_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "async_effect_gate.zen",
-        r#"
-main = (mode: Async) void { }
-"#,
-        "async effect gate",
-        "async effect gate should emit one Sync/Async feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_async_effect_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_dynamic_string_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "dynamic_string_gate.zen",
-        r#"
-main = (value: String) void { }
-"#,
-        "dynamic string gate",
-        "dynamic string gate should emit one allocator-backed text diagnostic",
-        "tests/fixtures/ir_json/diagnostics_dynamic_string_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_generic_dynamic_string_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "generic_dynamic_string_gate.zen",
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = (box: Box<String>) void { }
-"#,
-        "generic dynamic string gate",
-        "generic dynamic string gate should emit one allocator-backed text diagnostic",
-        "tests/fixtures/ir_json/diagnostics_generic_dynamic_string_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_range_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "range_gate.zen",
-        r#"
-main = () i32 {
-    1..3
-}
-"#,
-        "range gate",
-        "range gate should emit one feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_range_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_raise_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "raise_gate.zen",
-        r#"
-Result<T, E>:
-    Ok(T),
-    Err(E)
-
-main = () i32 {
-    value = Result<i32, StaticString>.Ok(1)
-    value.raise()
-}
-"#,
-        "raise gate",
-        "raise gate should emit one propagation feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_raise_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_await_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "await_gate.zen",
-        r#"
-Task<T>: {
-    value: T
-}
-
-main = () i32 {
-    task = Task<i32> { value: 1 }
-    task.await()
-}
-"#,
-        "await gate",
-        "await gate should emit one Sync/Async feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_await_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_actor_type_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "actor_type_gate.zen",
-        r#"
-main = (actor: Actor<i32>) void { }
-"#,
-        "actor framework type gate",
-        "actor framework type gate should emit one actor feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_actor_type_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_bare_actor_type_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "bare_actor_type_gate.zen",
-        r#"
-main = (actor: Actor) void { }
-"#,
-        "bare actor framework type gate",
-        "bare actor framework type gate should emit one actor feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_bare_actor_type_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_actor_import_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "actor_import_gate.zen",
-        r#"
-{ Actor } = @std.concurrency.actor.actor
-
-main = () void { }
-"#,
-        "actor framework import gate",
-        "actor framework import gate should emit one std actor feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_actor_import_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_allocator_import_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "allocator_import_gate.zen",
-        r#"
-{ Allocator } = @std.memory.allocator
-
-main = () void { }
-"#,
-        "allocator framework import gate",
-        "allocator framework import gate should emit one std allocator feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_allocator_import_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_async_runtime_import_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "async_runtime_import_gate.zen",
-        r#"
-{ Scheduler } = @std.concurrency.async.scheduler
-
-main = () void { }
-"#,
-        "async runtime import gate",
-        "async runtime import gate should emit one std async feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_async_runtime_import_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_sync_runtime_import_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "sync_runtime_import_gate.zen",
-        r#"
-{ Channel } = @std.concurrency.sync.channel
-
-main = () void { }
-"#,
-        "sync runtime import gate",
-        "sync runtime import gate should emit one std sync feature-gate diagnostic",
-        "tests/fixtures/ir_json/diagnostics_sync_runtime_import_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_async_intrinsic_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "async_intrinsic_gate.zen",
-        r#"
-main = () void {
-    @builtin.async_enqueue(1)
-}
-"#,
-        "async intrinsic gate",
-        "async intrinsic gate should emit one gated intrinsic diagnostic",
-        "tests/fixtures/ir_json/diagnostics_async_intrinsic_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_raw_allocate_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "raw_allocate_gate.zen",
-        r#"
-main = () void {
-    @builtin.raw_allocate(8)
-}
-"#,
-        "raw allocation gate",
-        "raw allocation gate should emit one gated intrinsic diagnostic",
-        "tests/fixtures/ir_json/diagnostics_raw_allocate_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_byte_memory_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "byte_memory_gate.zen",
-        r#"
-main = () void {
-    @builtin.memcpy(0, 0, 8)
-}
-"#,
-        "byte memory gate",
-        "byte memory gate should emit one gated intrinsic diagnostic",
-        "tests/fixtures/ir_json/diagnostics_byte_memory_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_raw_pointer_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "raw_pointer_gate.zen",
-        r#"
-main = () void {
-    @builtin.gep(0, 1)
-}
-"#,
-        "raw pointer gate",
-        "raw pointer gate should emit one gated intrinsic diagnostic",
-        "tests/fixtures/ir_json/diagnostics_raw_pointer_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_atomic_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "atomic_gate.zen",
-        r#"
-main = () void {
-    @builtin.atomic_load(0)
-}
-"#,
-        "atomic gate",
-        "atomic gate should emit one gated intrinsic diagnostic",
-        "tests/fixtures/ir_json/diagnostics_atomic_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_syscall_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "syscall_gate.zen",
-        r#"
-main = () void {
-    @builtin.syscall0(1)
-}
-"#,
-        "syscall gate",
-        "syscall gate should emit one gated intrinsic diagnostic",
-        "tests/fixtures/ir_json/diagnostics_syscall_gate.golden.json",
-    );
-}
-
-#[test]
-fn emit_json_diagnostics_type_match_gate_schema_matches_golden() {
-    assert_gate_diagnostics_golden(
-        "type_match_gate.zen",
-        r#"
-Point: {
-    x: i32,
-}
-
-main = () void {
-    @builtin.type_match<Point>()
-}
-"#,
-        "type match gate",
-        "type match gate should emit one gated intrinsic diagnostic",
-        "tests/fixtures/ir_json/diagnostics_type_match_gate.golden.json",
-    );
 }
